# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Trivela API
  version: 0.1.0
  description: |
    REST API for the Trivela campaign and rewards platform. Handles campaign metadata,
    health checks, and Soroban RPC configuration for the frontend.

    See the [API Migration Guide](../docs/API_MIGRATION.md) for v0 → v1 breaking changes,
    endpoint changelog, deprecation policy, and migration examples with curl snippets.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: https://api.trivela.example.com
    description: Production server
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /api/v1/admin/api-keys:
    get:
      summary: List API keys (metadata only)
      tags:
        - Admin
      operationId: listApiKeys
      responses:
        '200':
          description: API key list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiKeyMetadata'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a new API key
      tags:
        - Admin
      operationId: createApiKey
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
                expiresAt:
                  type: string
                  format: date-time
                  nullable: true
                orgId:
                  type: string
                  description: Scope the key to an organization.
                scopes:
                  type: array
                  items:
                    $ref: '#/components/schemas/ApiKeyScope'
                  description: Defaults to the full legacy scope set if omitted.
                rateTier:
                  $ref: '#/components/schemas/ApiKeyRateTier'
                  description: Defaults to `standard` if omitted.
      responses:
        '201':
          description: API key created (raw key returned once)
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
                  metadata:
                    $ref: '#/components/schemas/ApiKeyMetadata'
  /api/v1/admin/api-keys/{id}:
    delete:
      summary: Revoke an API key
      tags:
        - Admin
      operationId: revokeApiKey
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Key revoked
        '404':
          description: Key not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/admin/api-keys/{id}/rotate:
    put:
      summary: Rotate an API key (revoke old, issue new)
      tags:
        - Admin
      operationId: rotateApiKey
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: New key issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
                  metadata:
                    $ref: '#/components/schemas/ApiKeyMetadata'
  /api/v1/admin/api-keys/{id}/rate-tier:
    put:
      summary: Update an API key's rate tier
      description: >-
        Changes the enforced rate-limit tier in place, without rotating (and
        thus invalidating) the caller's existing credential.
      tags:
        - Admin
      operationId: updateApiKeyRateTier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - rateTier
              properties:
                rateTier:
                  $ref: '#/components/schemas/ApiKeyRateTier'
      responses:
        '200':
          description: Rate tier updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  metadata:
                    $ref: '#/components/schemas/ApiKeyMetadata'
        '400':
          description: Invalid rate tier value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '404':
          description: Key not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for write operations
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication
  schemas:
...
    ApiKeyMetadata:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        orgId:
          type: string
          nullable: true
          description: Organization the key is scoped to, if any.
        scopes:
          type: array
          items:
            $ref: '#/components/schemas/ApiKeyScope'
          description: Granular permissions granted to the key.
        rateTier:
          $ref: '#/components/schemas/ApiKeyRateTier'
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
        active:
          type: boolean
    ApiKeyScope:
      type: string
      enum: [campaigns:read, campaigns:write, allowlist:write, admin]
    ApiKeyRateTier:
      type: string
      enum: [standard, pro, enterprise]
      description: >-
        Per-key rate-limit tier enforced by the rate limiter (#924). `standard`
        is 60 req/min, `pro` is 300 req/min, `enterprise` is 1000 req/min.
...
    Error:
      type: object
      required:
        - error
        - code
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
    ValidationError:
      type: object
      required:
        - error
        - code
        - details
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: array
          items:
            type: string
          description: Validation error details
...